Extract Tooltip question mark component
